Refactored CategoryChart tests
CategoryChart tests are now on its own file
